// packages/astro/dist/assets/types.d.ts
export type ImageInputFormat = (typeof VALID_INPUT_FORMATS)[number];
// ...
declare const isESMImport: unique symbol;
export type OmitBrand<T> = Omit<T, typeof isESMImport>;
/**
 * Type returned by ESM imports of images
 */
export type ImageMetadata = {
	src: string;
	width: number;
	height: number;
	format: ImageInputFormat;
	orientation?: number;
	[isESMImport]?: true;
};